Add autochip/kick to Move Primitive
